Revert "Create multiple components for data management (#106)"
This reverts commit 33be492be641543ea154ab28177fb02cd7fbd0ba.
