--- src/interfaces/SerialDataRetriever.ts
@@ -1,4 +1,0 @@
-export default interface SerialDataRetriever {
-  start: () => void;
-  stop: () => void;
-}
